requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.2
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.6
django-filter==21.1
django-ipware==4.0.2
django-session-timeout==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
Pillow==8.4.0
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
